Fixes form refactoring
In func.py and op.py there was a circular import that has been removed
In immutable/heap.py there was a logic bug introduced by me (fixed)
In iters.py the .uniform module is used to make P2 and P3 work the same
way. can't make this PEP8 compilent. Have to `import *`