Move python 3.6 and pypy3 environments to their own travis image.
Python 3.6 failed silently: https://travis-ci.org/medecau/fn.py/jobs/222589293#L150
Python 3.6 is not available in python 3.5 base image and 3.5 is not available in the 3.6 image.

Pypy3 failed catastrophically
The version of pypy3 currently provided by travis is old and is no longer supported by the most recent version of pip.
The solution, for now, is to use the pypy3 image.

Related issues:
https://github.com/travis-ci/travis-ci/issues/4794
https://github.com/travis-ci/travis-ci/issues/4990
https://github.com/travis-ci/travis-ci/issues/7146